+++ audit/forgerock-audit-handler-jms/src/main/java/org/forgerock/audit/handlers/jms/JmsAuditEventHandlerConfiguration.java
@@ -16,11 +16,18 @@
 
 package org.forgerock.audit.handlers.jms;
 
+import java.util.Collections;
+import java.util.Map;
+import javax.jms.ConnectionFactory;
 import javax.jms.JMSContext;
+import javax.jms.Topic;
+import javax.naming.InitialContext;
+
+import org.forgerock.audit.events.handlers.EventHandlerConfiguration;
+import org.forgerock.util.Reject;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyDescription;
-import org.forgerock.audit.events.handlers.EventHandlerConfiguration;
 
 /**
  * Configuration object for the {@link JmsAuditEventHandler}.
@@ -31,27 +38,20 @@ import org.forgerock.audit.events.handlers.EventHandlerConfiguration;
  * {
  *     "name" : "jms",
  *     "topics": [ "access", "activity", "config", "authentication" ],
- *     "providerUrl" : "tcp://localhost:61616",
- *     "jmsTopic" : "audit",
- *     "initialContextFactory" : "org.apache.activemq.jndi.ActiveMQInitialContextFactory",
  *     "deliveryMode" : "NON_PERSISTENT",
- *     "sessionMode" : "AUTO"
+ *     "sessionMode" : "AUTO",
+ *     "jndi" :  {
+ *          "contextProperties" : {
+ *              "initialContextFactory" : "org.apache.activemq.jndi.ActiveMQInitialContextFactory",
+ *              "providerUrl" : "tcp://localhost:61616"
+ *          },
+ *          "topicName" : "audit",
+ *          "connectionFactoryName" : "connectionFactory"
+ *     }
  * }
  * </pre>
  */
 public class JmsAuditEventHandlerConfiguration extends EventHandlerConfiguration {
-
-    @JsonProperty(required = true)
-    @JsonPropertyDescription("audit.handlers.jms.providerUrl")
-    private String providerUrl;
-
-    @JsonProperty(required = true)
-    @JsonPropertyDescription("audit.handlers.jms.jmsTopic")
-    private String jmsTopic;
-
-    @JsonProperty(required = true)
-    @JsonPropertyDescription("audit.handlers.jms.initialContextFactory")
-    private String initialContextFactory;
 
     @JsonProperty(required = true)
     @JsonPropertyDescription("audit.handlers.jms.deliveryMode")
@@ -61,65 +61,11 @@ public class JmsAuditEventHandlerConfiguration extends EventHandlerConfiguration
     @JsonPropertyDescription("audit.handlers.jms.sessionMode")
     private SessionModeConfig sessionMode;
 
-    @JsonPropertyDescription("audit.handlers.jms.batchEnabled")
-    private boolean batchEnabled;
+    @JsonPropertyDescription("audit.handlers.jms.batch")
+    private BatchPublisherConfiguration batch = new BatchPublisherConfiguration();
 
-    @JsonPropertyDescription("audit.handlers.jms.batchConfiguration")
-    private BatchPublisherConfiguration batchConfiguration;
-
-    /**
-     * Returns the JMS provider connection url utilized to connect to the JMS broker.
-     *
-     * @return the JMS provider connection url utilized to connect to the JMS broker.
-     */
-    public String getProviderUrl() {
-        return providerUrl;
-    }
-
-    /**
-     * Sets the JMS provider connection url utilized to connect to the JMS broker.
-     *
-     * @param providerUrl the connection url
-     */
-    public void setProviderUrl(String providerUrl) {
-        this.providerUrl = providerUrl;
-    }
-
-    /**
-     * Returns the JMS topic to which messages will be published.  Do not confuse this with Audit Topics.
-     *
-     * @return The JMS topic to which messages will be published.
-     */
-    public String getJmsTopic() {
-        return jmsTopic;
-    }
-
-    /**
-     * Sets the JMS topic for which the messages will be published on.
-     *
-     * @param jmsTopic the JMS topic
-     */
-    public void setJmsTopic(String jmsTopic) {
-        this.jmsTopic = jmsTopic;
-    }
-
-    /**
-     * Returns the classname that implements the JMS connection Factory.
-     *
-     * @return the classname that implements the JMS connection Factory.
-     */
-    public String getInitialContextFactory() {
-        return initialContextFactory;
-    }
-
-    /**
-     * Sets the classname that implements the JMS Connection Factory.
-     *
-     * @param initialContextFactory the JMS Connection Factory classname
-     */
-    public void setInitialContextFactory(String initialContextFactory) {
-        this.initialContextFactory = initialContextFactory;
-    }
+    @JsonPropertyDescription("audit.handlers.jms.jndi")
+    private JndiConfiguration jndi = new JndiConfiguration();
 
     /**
      * Returns the delivery mode configuration that should be used when publishing the JMS messages.
@@ -160,38 +106,111 @@ public class JmsAuditEventHandlerConfiguration extends EventHandlerConfiguration
     }
 
     /**
-     * Returns true if handling of audit events should be done in batches.
-     *
-     * @return true if handling of audit events should be done in batches.
-     */
-    public boolean isBatchEnabled() {
-        return batchEnabled;
-    }
-
-    /**
-     * sets if handling of audit events should be done in batches.
-     *
-     * @param batchEnabled true if handling of audit events should be done in batches.
-     */
-    public void setBatchEnabled(boolean batchEnabled) {
-        this.batchEnabled = batchEnabled;
-    }
-
-    /**
      * Returns the configuration used to initialize the batch publisher.
      *
      * @return the configuration used to initialize the batch publisher.
      */
-    public BatchPublisherConfiguration getBatchConfiguration() {
-        return batchConfiguration;
+    public BatchPublisherConfiguration getBatch() {
+        return batch;
     }
 
     /**
      * Sets the configuration used to initialize the batch publisher.
      *
-     * @param batchConfiguration the configuration used to initialize the batch publisher.
+     * @param batch the configuration used to initialize the batch publisher.
      */
-    public void setBatchConfiguration(BatchPublisherConfiguration batchConfiguration) {
-        this.batchConfiguration = batchConfiguration;
+    public void setBatch(BatchPublisherConfiguration batch) {
+        this.batch = batch;
+    }
+
+    /**
+     * Gets the {@link JndiConfiguration}.
+     * @return The {@link JndiConfiguration}.
+     */
+    public JndiConfiguration getJndi() {
+        return jndi;
+    }
+
+    /**
+     * Sets the {@link JndiConfiguration}.
+     * @param jndi The {@link JndiConfiguration}
+     */
+    public void setJndi(JndiConfiguration jndi) {
+        this.jndi = jndi;
+    }
+
+
+    /**
+     * Stores the JNDI context properties and lookup names.
+     */
+    public static class JndiConfiguration {
+
+        @JsonPropertyDescription("audit.handlers.jms.contextProperties")
+        private Map<String, String> contextProperties = Collections.emptyMap();
+
+        @JsonPropertyDescription("audit.handlers.jms.topicName")
+        private String topicName = "audit";
+
+        @JsonPropertyDescription("audit.handlers.jms.connectionFactoryName")
+        private String connectionFactoryName = "connectionFactory";
+
+        /**
+         * Gets the Jndi {@link InitialContext} properties.
+         * @return The {@link InitialContext} properties.
+         */
+        public Map<String, String> getContextProperties() {
+            return contextProperties;
+        }
+
+        /**
+         * Sets the Jndi {@link InitialContext} properties.
+         * @param contextProperties The {@link InitialContext} properties.
+         */
+        public void setContextProperties(Map<String, String> contextProperties) {
+            Reject.ifNull(contextProperties, "The jndi context properties can't be null");
+            this.contextProperties = Collections.unmodifiableMap(contextProperties);
+        }
+
+        /**
+         * Returns the jndi lookup name for the JMS {@link Topic} to which messages will be published.
+         * Do not confuse this with Audit Topics.
+         * @see InitialContext#lookup(String)
+         *
+         * @return The jndi lookup name for the JMS {@link Topic} to which messages will be published.
+         */
+        public String getTopicName() {
+            return topicName;
+        }
+
+        /**
+         * Sets the jndi lookup name for the JMS {@link Topic} for which the messages will be published on.
+         * @see InitialContext#lookup(String)
+         *
+         * @param jmsTopicName The jndi lookup name for the JMS {@link Topic}.
+         */
+        public void setJmsTopicName(String jmsTopicName) {
+            this.topicName = jmsTopicName;
+        }
+
+        /**
+         * Returns the jndi lookup name for the JMS {@link ConnectionFactory} to which messages will be published.
+         * Do not confuse this with Audit Topics.
+         * @see InitialContext#lookup(String)
+         *
+         * @return The jndi lookup name for the JMS {@link ConnectionFactory} to which messages will be published.
+         */
+        public String getConnectionFactoryName() {
+            return connectionFactoryName;
+        }
+
+        /**
+         * Sets the jndi lookup name for the JMS {@link ConnectionFactory} for which the messages will be published on.
+         * @see InitialContext#lookup(String)
+         *
+         * @param jmsConnectionFactoryName The jndi lookup name for the JMS {@link ConnectionFactory}.
+         */
+        public void setJmsConnectionFactoryName(String jmsConnectionFactoryName) {
+            this.connectionFactoryName = jmsConnectionFactoryName;
+        }
     }
 }